Sun bug 6359959: xmag overconsumes CPU due to its selection box
<http://bugs.opensolaris.org/bugdatabase/view_bug.do?bug_id=6359959>

The fix for Sun bug 1166559 changed the HLINTERVAL from 10 to 0.5 to fix
a problem that the selection box wasn't being seen. The problem is that
this parameter is used in an XtAppAddTimeOut(), and is supposed to be an
unsigned long. Thus, the 0.5 presumably turns into a 0 ms timeout. Whatever
effect that is being seen is due to seeing a rectangle drawn as fast as
the poor little X server can do it.