Properly handle multiple depth windows.
This patch will not cause X Errors when magnifying windows with portions
offscreen, or windows with a dimension smaller than the default 64x64.

A common case where this patch is required is when Composite is enabled,
the default depth has a value other than 24, and some application is
using the Composite argb visual; example is WindowMaker, that will choose
the "best" available visual, unless the visual-id option is used.